Refactored message reaction code into /utility/postjob-reaction.js
